Move out enums to a new file
