+++ frontend/src/@types/GroceryItemContext.d.ts
@@ -16,18 +16,3 @@ export interface GroceryItemResponse {
     message: string,
     status: number
 }
-
-export const Unit: string[] = ['pc', 'kg', 'lb', 'L'];
-
-export const Category: string[] = [
-    'Meat',
-    'Seafood',
-    'Vegetable',
-    'Fruit',
-    'Snack',
-    'Dairy/Egg',
-    'Frozen',
-    'Condiment/Seasoning',
-    'Fermented',
-    'Beverage'
-]
