+++ frontend/src/enums.tsx
@@ -1,0 +1,14 @@
+export const Unit: string[] = ['pc', 'kg', 'lb', 'L'];
+
+export const Category: string[] = [
+    'Meat',
+    'Seafood',
+    'Vegetable',
+    'Fruit',
+    'Snack',
+    'Dairy/Egg',
+    'Frozen',
+    'Condiment/Seasoning',
+    'Fermented',
+    'Beverage'
+]
